refactor: consolidate `addMessageToDiv` and `handleResponse` into `handleChat`
refactor: rename `handleChat` function to `handleChatRequest` to avoid confusion withh filename
